Use consistent scalar types
Address compiler diagnostics for implicit floating-point conversions
an integer conversions. Fix Fortran unit test error messages.

+++ include/heffte_fft3d.h
@@ -637,13 +637,13 @@ private:
                 std::is_same<backend_tag, backend::fftw_sin>::value) {
                 scale_factor /= 8.0;
             }else if (std::is_same<backend_tag, backend::fftw_cos1>::value) {
-                scale_factor = 1.0 / (8.0 * (plan.fft_sizes[0] - 1) * (plan.fft_sizes[1] - 1) * (plan.fft_sizes[2] - 1));
+                scale_factor = 1.0 / (8.0 * (double)((plan.fft_sizes[0] - 1) * (plan.fft_sizes[1] - 1) * (plan.fft_sizes[2] - 1)));
             }else if (std::is_same<backend_tag, backend::fftw_sin1>::value) {
-                scale_factor = 1.0 / (8.0 * (plan.fft_sizes[0] + 1) * (plan.fft_sizes[1] + 1) * (plan.fft_sizes[2] + 1));
+                scale_factor = 1.0 / (8.0 * (double)((plan.fft_sizes[0] + 1) * (plan.fft_sizes[1] + 1) * (plan.fft_sizes[2] + 1)));
             }else if(std::is_same<backend_tag, backend::cufft_cos1>::value or
                      std::is_same<backend_tag, backend::rocfft_cos1>::value or
                      std::is_same<backend_tag, backend::stock_cos1>::value){
-                scale_factor = 1.0 / (64.0 * (plan.fft_sizes[0] - 1) * (plan.fft_sizes[1] - 1) * (plan.fft_sizes[2] - 1));
+                scale_factor = 1.0 / (64.0 * (double)((plan.fft_sizes[0] - 1) * (plan.fft_sizes[1] - 1) * (plan.fft_sizes[2] - 1)));
             }else{
                 scale_factor /= 64.0;
             }
